refactor(contacts): centralize flow package exports
