+++ apps/ledger-live-mobile/src/mvvm/features/Contacts/types.ts
@@ -1,1 +1,1 @@
-export type { ContactsFeatureConfig as ContactsEntryConfig } from "@features/flow-contacts/featureFlags";
+export type { ContactsFeatureConfig as ContactsEntryConfig } from "@features/flow-contacts";
