+++ apps/ledger-live-mobile/src/screens/Settings/Debug/Debugging/Contacts/constants.ts
@@ -1,7 +1,7 @@
 import {
   CONTACTS_FEATURE_FLAG_KEYS,
   DEFAULT_ELIGIBLE_ADDRESS_FAMILIES,
-} from "@features/flow-contacts/featureFlags";
+} from "@features/flow-contacts";
 
 export const CONTACTS_FLAG = CONTACTS_FEATURE_FLAG_KEYS.mobile;
