+++ features/flow/contacts/README.md
@@ -11,6 +11,12 @@ Shared Contacts flow package for Desktop and Mobile.
 
 App layers own routing, screen composition, i18n, and analytics.
 
+## Public API
+
+Consume the package from `@features/flow-contacts`. The root entry point resolves to the
+appropriate Web or React Native API. Folders under `src/` are internal implementation details
+and are not exported as package subpaths.
+
 ## Structure
 
 ```
