+++ features/flow/contacts/src/index.native.ts
@@ -1,0 +1,3 @@
+export * from "./featureFlags";
+export { useContactsMeContact } from "./hooks";
+export * from "./list";
