+++ features/flow/contacts/src/index.ts
@@ -1,3 +1,7 @@
 export * from "./featureFlags";
 export { useContactsMeContact } from "./hooks";
 export * from "./list";
+export { ContactsButton } from "./components/ContactsButton";
+export type { ContactsButtonProps } from "./components/ContactsButton";
+export { ContactsAddContactButton, ContactsMeItem } from "./components/ContactsPage";
+export type { ContactsAddContactButtonProps, ContactsMeItemProps } from "./components/ContactsPage";
