+++ features/flow/contacts/src/list/internals/getContactInitial.test.ts
@@ -4,7 +4,7 @@ describe("getContactInitial", () => {
   it("returns the first letter and its combining marks in uppercase", () => {
     expect(getContactInitial("olive")).toBe("O");
     expect(getContactInitial("eleonore")).toBe("E");
-    expect(getContactInitial("e\u0301lodie")).toBe("E\u0301");
+    expect(getContactInitial("élodie")).toBe("É");
     expect(getContactInitial("Алексей")).toBe("А");
     expect(getContactInitial("مريم")).toBe("م");
   });
