+++ features/flow/contacts/src/list/internals/getContactInitial.ts
