+++ features/flow/contacts/src/list/internals/index.ts
@@ -1,0 +1,1 @@
+export { getContactInitial } from "./getContactInitial";
